fixed the outstanding issues
